chore: 커리큘럼 조회 api CurriculumController로 이동
#119
