+++ src/main/java/doore/study/api/StudyController.java
@@ -4,11 +4,8 @@ import doore.study.application.StudyCommandService;
 import doore.study.application.StudyQueryService;
 import doore.study.application.dto.request.StudyCreateRequest;
 import doore.study.application.dto.request.StudyUpdateRequest;
-import doore.study.application.dto.response.CurriculumItemResponse;
-import doore.study.application.dto.response.PersonalCurriculumItemResponse;
 import doore.study.application.dto.response.StudyResponse;
 import doore.study.application.dto.response.StudySimpleResponse;
-import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -75,19 +72,5 @@ public class StudyController {
     public ResponseEntity<List<StudySimpleResponse>> getMyStudies(@PathVariable final Long memberId) {
         // TODO: 3/22/24 토큰의 주인과 회원아이디가 같은지 검증
         return ResponseEntity.ok(studyQueryService.findMyStudies(memberId));
-    }
-
-    @GetMapping("studies/{studyId}/curriculums/all")
-    public ResponseEntity<List<CurriculumItemResponse>> getCurriculums(@PathVariable Long studyId) {
-        List<CurriculumItemResponse> responses = studyQueryService.getCurriculums(studyId);
-        return ResponseEntity.ok(responses);
-    }
-
-    @GetMapping("studies/{studyId}/curriculums")
-    public ResponseEntity<List<PersonalCurriculumItemResponse>> getMyCurriculum(@PathVariable Long studyId,
-                                                                                HttpServletRequest request) {
-        String memberId = request.getHeader("Authorization");
-        List<PersonalCurriculumItemResponse> response = studyQueryService.getMyCurriculum(studyId, Long.parseLong(memberId));
-        return ResponseEntity.ok(response);
     }
 }
