+++ src/main/java/doore/study/application/StudyQueryService.java
@@ -8,14 +8,9 @@ import static java.util.stream.Collectors.groupingBy;
 import doore.crop.domain.Crop;
 import doore.crop.domain.repository.CropRepository;
 import doore.crop.exception.CropException;
-import doore.study.application.dto.response.CurriculumItemResponse;
-import doore.study.application.dto.response.PersonalCurriculumItemResponse;
 import doore.study.application.dto.response.StudyResponse;
 import doore.study.application.dto.response.StudySimpleResponse;
-import doore.study.domain.ParticipantCurriculumItem;
 import doore.study.domain.Study;
-import doore.study.domain.repository.CurriculumItemRepository;
-import doore.study.domain.repository.ParticipantCurriculumItemRepository;
 import doore.study.domain.repository.StudyRepository;
 import doore.study.exception.StudyException;
 import doore.study.persistence.StudyDao;
@@ -40,8 +35,6 @@ public class StudyQueryService {
     private final TeamRepository teamRepository;
     private final CropRepository cropRepository;
     private final StudyDao studyDao;
-    private final CurriculumItemRepository curriculumItemRepository;
-    private final ParticipantCurriculumItemRepository participantCurriculumItemRepository;
 
     public StudyResponse findStudyById(Long studyId) {
         Study study = studyRepository.findById(studyId).orElseThrow(() -> new StudyException(NOT_FOUND_STUDY));
@@ -50,15 +43,6 @@ public class StudyQueryService {
         final Crop crop = cropRepository.findById(study.getCropId())
                 .orElseThrow(() -> new CropException(NOT_FOUND_CROP));
         return StudyResponse.of(study, team, crop);
-    }
-
-    public List<CurriculumItemResponse> getCurriculums(Long studyId) {
-        return CurriculumItemResponse.from(curriculumItemRepository.findAllByStudyId(studyId));
-    }
-
-    public List<PersonalCurriculumItemResponse> getMyCurriculum(Long studyId, Long memberId) {
-        List<ParticipantCurriculumItem> participantCurriculumItems = participantCurriculumItemRepository.findAllByStudyIdAndMemberId(studyId, memberId);
-        return participantCurriculumItems.stream().map(PersonalCurriculumItemResponse::from).toList();
     }
 
     public List<StudySimpleResponse> findMyStudies(final Long memberId) {
